Update tokenizer.py to include stopwords support
Create a stopWords dictionary from nltk stopwords.
Every line split by newline at line 15 is split by the RegexpTokenizer into 'tokenized_line'.
Every word in the 'tokenized_line' that is not in 'stopWords' is appended to a temporary 'real_line' list. 
If the 'real_list' contains items, it is appended to the list 'lines'. 
'lines' is returned by the function